Only one helper module with less functions needed
