+++ pynest/examples/EI_clustered_network/helper.py
@@ -23,16 +23,17 @@
 ------------------------------------------------
 
 Helper functions to calculate synaptic weights and first passage times for exponential synapses to construct
-random balanced networks.
+random balanced networks and plot raster plot with color groups.
 """
 
+import matplotlib.pyplot as plt
 import numpy as np
 
 small = 1e-10
 
 
-def max_psp_exp(tau_m, tau_syn, c_m=1.0, e_l=0.0):
-    """Maximum psp amplitude for exponential synapses and unit J.
+def postsynaptic_current_to_potential(tau_m, tau_syn, c_m=1.0, e_l=0.0):
+    """Maximum post-synaptic potential amplitude for exponential synapses and a synaptic efficacy J of 1 pA.
 
     Parameters
     ----------
@@ -48,15 +49,15 @@ def max_psp_exp(tau_m, tau_syn, c_m=1.0, e_l=0.0):
     Returns
     -------
     float
-        maximum psp amplitude [mV]
+        maximum psp amplitude (for a 1 pA current spike) [mV]
     """
-    tmax = np.log(tau_syn / tau_m) / (1 / tau_m - 1 / tau_syn)
-    B = tau_m * tau_syn / c_m / (tau_syn - tau_m)
-    return (e_l - B) * np.exp(-tmax / tau_m) + B * np.exp(-tmax / tau_syn)
+    tmax = np.log(tau_syn / tau_m) / (1 / tau_m - 1 / tau_syn)  # time of maximum deflection of the psp  [ms]
+    pre = tau_m * tau_syn / c_m / (tau_syn - tau_m)  # we assume here the current spike is 1 pA, otherwise [mV/pA]
+    return (e_l - pre) * np.exp(-tmax / tau_m) + pre * np.exp(-tmax / tau_syn)
 
 
-def calc_js(params):
-    """Calculate synaptic weights for exponential synapses before clustering.
+def calculate_RBN_weights(params):
+    """Calculate synaptic weights for exponential synapses before clustering, for a random balanced network.
 
     Parameters
     ----------
@@ -71,41 +72,34 @@ def calc_js(params):
     N_E = params.get("N_E")  # excitatory units
     N_I = params.get("N_I")  # inhibitory units
     N = N_E + N_I  # total units
-    ps = params.get("ps")  # connection probs
-    ge = params.get("ge")
-    gi = params.get("gi")
+    baseline_conn_prob = params.get("baseline_conn_prob")  # connection probs
+    gei = params.get("gei")
+    gii = params.get("gii")
     gie = params.get("gie")
     V_th_E = params.get("V_th_E")  # threshold voltage
     V_th_I = params.get("V_th_I")
     tau_E = params.get("tau_E")
     tau_I = params.get("tau_I")
     E_L = params.get("E_L")
-    neuron_type = params.get("neuron_type")
-    if ("iaf_psc_exp" in neuron_type) or ("gif_psc_exp" in neuron_type):
-        tau_syn_ex = params.get("tau_syn_ex")
-        tau_syn_in = params.get("tau_syn_in")
-        amp_EE = max_psp_exp(tau_E, tau_syn_ex)
-        amp_EI = max_psp_exp(tau_E, tau_syn_in)
-        amp_IE = max_psp_exp(tau_I, tau_syn_ex)
-        amp_II = max_psp_exp(tau_I, tau_syn_in)
-    else:
-        amp_EE = 1.0
-        amp_EI = 1.0
-        amp_IE = 1.0
-        amp_II = 1.0
+    tau_syn_ex = params.get("tau_syn_ex")
+    tau_syn_in = params.get("tau_syn_in")
+    amp_EE = postsynaptic_current_to_potential(tau_E, tau_syn_ex)
+    amp_EI = postsynaptic_current_to_potential(tau_E, tau_syn_in)
+    amp_IE = postsynaptic_current_to_potential(tau_I, tau_syn_ex)
+    amp_II = postsynaptic_current_to_potential(tau_I, tau_syn_in)
 
     js = np.zeros((2, 2))
-    K_EE = N_E * ps[0, 0]
+    K_EE = N_E * baseline_conn_prob[0, 0]
     js[0, 0] = (V_th_E - E_L) * (K_EE**-0.5) * N**0.5 / amp_EE
-    js[0, 1] = -ge * js[0, 0] * ps[0, 0] * N_E * amp_EE / (ps[0, 1] * N_I * amp_EI)
-    K_IE = N_E * ps[1, 0]
+    js[0, 1] = -gei * js[0, 0] * baseline_conn_prob[0, 0] * N_E * amp_EE / (baseline_conn_prob[0, 1] * N_I * amp_EI)
+    K_IE = N_E * baseline_conn_prob[1, 0]
     js[1, 0] = gie * (V_th_I - E_L) * (K_IE**-0.5) * N**0.5 / amp_IE
-    js[1, 1] = -gi * js[1, 0] * ps[1, 0] * N_E * amp_IE / (ps[1, 1] * N_I * amp_II)
+    js[1, 1] = -gii * js[1, 0] * baseline_conn_prob[1, 0] * N_E * amp_IE / (baseline_conn_prob[1, 1] * N_I * amp_II)
     return js
 
 
-def fpt(tau_m, e_l, i_e, c_m, vtarget, vstart):
-    """Calculate first pasage time between Vstart and Vtarget.
+def rheobase_current(tau_m, e_l, v_th, c_m):
+    """Rheobase current for a given membrane time constant and resting potential.
 
     Parameters
     ----------
@@ -113,56 +107,60 @@ def fpt(tau_m, e_l, i_e, c_m, vtarget, vstart):
         Membrane time constant [ms]
     e_l: float
         Resting potential [mV]
-    i_e: float
-        external current [pA]
+    v_th: float
+        Threshold potential [mV]
     c_m: float
         Membrane capacity [pF]
-    vtarget: float
-        target voltage [mV]
-    vstart: float
-        start voltage [mV]
 
     Returns
     -------
     float
-        first passage time [ms]
+        rheobase current [pA]
     """
-    inner = (vtarget - e_l - tau_m * i_e / c_m) / (
-        vstart - e_l - tau_m * i_e / c_m + small
-    )
-    if inner < 0:
-        return np.nan
-    else:
-        return -tau_m * np.log(inner)
+    return (v_th - e_l) * c_m / tau_m
 
 
-def v_fpt(tau_m, e_l, i_e, c_m, ttarget, vtarget, t_ref):
-    """Initial voltage to obtain a certain first passage time.
+def raster_plot(spiketimes, tlim=None, colorgroups=None, ax=None, markersize=0.5):
+    """Raster plot of spiketimes.
+
+    Plots raster plot of spiketimes withing given time limits and colors neurons according to colorgroups.
 
     Parameters
     ----------
-    tau_m: float
-        Membrane time constant [ms]
-    e_l: float
-        Resting potential [mV]
-    i_e: float
-        external current [pA]
-    c_m: float
-        Membrane capacity [pF]
-    ttarget: float
-        target first passage time [ms]
-    vtarget: float
-        target voltage [mV]
-    t_ref: float
-        refractory period [ms]
+    spiketimes: ndarray
+        2D array [2xN_Spikes] of spiketimes with spiketimes in row 0 and neuron IDs in row 1.
+    tlim: list of floats (optional)
+        Time limits of plot: [tmin, tmax], if None: [min(spiketimes), max(spiketimes)]
+    colorgroups: list of tuples (optional)
+        Each element is a tuple in the format (color, start_neuron_ID, stop_neuron_ID]) for coloring neurons in
+        given range, if None is given all neurons are black.
+    ax: axis object (optional)
+        If None a new figure is created, else the plot is added to the given axis.
+    markersize: float (optional)
+        Size of markers. (default: 0.5)
 
     Returns
     -------
-    float
-        initial voltage [mV]
+    axis object
+        Axis object with raster plot.
     """
-    return (
-        (vtarget - e_l - tau_m * i_e / c_m) * np.exp(ttarget / tau_m)
-        + e_l
-        + tau_m * i_e / c_m
-    )
+    if ax is None:
+        fig, ax = plt.subplots()
+    if tlim is None:
+        tlim = [min(spiketimes[0]), max(spiketimes[0])]
+    if colorgroups is None:
+        colorgroups = [("k", 0, max(spiketimes[1]))]
+    for color, start, stop in colorgroups:
+        ax.plot(
+            spiketimes[0][np.logical_and(spiketimes[1] >= start, spiketimes[1] < stop)],
+            spiketimes[1][np.logical_and(spiketimes[1] >= start, spiketimes[1] < stop)],
+            color=color,
+            marker=".",
+            linestyle="None",
+            markersize=markersize,
+        )
+    ax.set_xlim(tlim)
+    ax.set_ylim([0, max(spiketimes[1])])
+    ax.set_xlabel("Time [ms]")
+    ax.set_ylabel("Neuron ID")
+    return ax
